+++ Citadel/src/utopia/citadel/database/access/single/description/DescriptionLinkAccess.scala
@@ -2,12 +2,12 @@ package utopia.citadel.database.access.single.description
 
 import utopia.citadel.database.factory.description.DescriptionLinkFactory
 import utopia.citadel.database.model.description.{DescriptionLinkModelFactory, DescriptionModel}
-import utopia.citadel.model.enumeration.StandardDescriptionRoleId
+import utopia.citadel.model.enumeration.CitadelDescriptionRole.Name
+import utopia.metropolis.model.enumeration.DescriptionRoleIdWrapper
 import utopia.metropolis.model.stored.description.DescriptionLink
 import utopia.vault.database.Connection
 import utopia.vault.model.immutable.Storable
-import utopia.vault.nosql.access.single.model.SingleModelAccess
-import utopia.vault.sql.{Select, Where}
+import utopia.vault.nosql.access.single.model.{SingleModelAccess, SingleRowModelAccess}
 
 /**
   * A common trait for individual description link access points
@@ -49,9 +49,17 @@ trait DescriptionLinkAccess extends SingleModelAccess[DescriptionLink]
 	 * @return That description (as text) of this item
 	 */
 	def apply(languageId: Int, roleId: Int)(implicit connection: Connection) =
-		connection(Select(target, descriptionModel.textColumn) +
-			Where(mergeCondition(descriptionModel.withLanguageId(languageId).withRoleId(roleId))))
-			.firstValue.string
+		findColumn(descriptionModel.textColumn,
+			descriptionModel.withLanguageId(languageId).withRoleId(roleId).toCondition).string
+	/**
+	  * Reads a single description text
+	  * @param languageId Id of the language the description should be in
+	  * @param role the role the description has
+	  * @param connection Implicit DB Connection
+	  * @return That description (as text) of this item
+	  */
+	def apply(languageId: Int, role: DescriptionRoleIdWrapper)(implicit connection: Connection): Option[String] =
+		apply(languageId, role.id)
 	
 	/**
 	  * @param languageId Id of targeted language
@@ -62,16 +70,15 @@ trait DescriptionLinkAccess extends SingleModelAccess[DescriptionLink]
 	
 	// NESTED	-------------------------
 	
-	class DescriptionInLanguage(val languageId: Int) extends SingleModelAccess[DescriptionLink]
+	class DescriptionInLanguage(val languageId: Int) extends SingleRowModelAccess[DescriptionLink]
 	{
 		// COMPUTED ---------------------
 		
 		/**
 		 * @param connection Implicit DB Connection
-		 * @return Link to the name description of this item
+		 * @return Name of this item in targeted language
 		 */
-		def name(implicit connection: Connection) =
-			forRoleWithId(StandardDescriptionRoleId.name)
+		def name(implicit connection: Connection) = apply(Name)
 		
 		
 		// IMPLEMENTED	-----------------
@@ -85,11 +92,27 @@ trait DescriptionLinkAccess extends SingleModelAccess[DescriptionLink]
 		// OTHER	---------------------
 		
 		/**
+		  * @param role targeted description role
+		  * @param connection Implicit connection
+		  * @return Description of that role in targeted language, as text
+		  */
+		def apply(role: DescriptionRoleIdWrapper)(implicit connection: Connection) =
+			textForRoleWithId(role.id)
+		
+		/**
 		  * @param roleId     Targeted description role's id
 		  * @param connection Db Connection
 		  * @return Description for that role for this item in targeted language
 		  */
 		def forRoleWithId(roleId: Int)(implicit connection: Connection): Option[DescriptionLink] =
-			read(Some(descriptionModel.withRoleId(roleId).toCondition))
+			find(descriptionModel.withRoleId(roleId).toCondition)
+		
+		/**
+		  * @param roleId Id of the targeted description role
+		  * @param connection Implicit connection
+		  * @return Description of that role in targeted language, as text
+		  */
+		def textForRoleWithId(roleId: Int)(implicit connection: Connection) =
+			findColumn(descriptionModel.textColumn, descriptionModel.withRoleId(roleId).toCondition).string
 	}
 }