+++ Citadel/src/utopia/citadel/database/access/single/organization/DbOrganization.scala
@@ -6,7 +6,7 @@ import utopia.citadel.database.access.many.organization.{DbUserRoles, Invitation
 import utopia.citadel.database.access.many.user.DbUsers
 import utopia.citadel.database.factory.organization.{MembershipFactory, MembershipWithRolesFactory}
 import utopia.citadel.database.model.organization.{DeletionModel, MemberRoleModel, MembershipModel, OrganizationModel}
-import utopia.citadel.model.enumeration.StandardDescriptionRoleId
+import utopia.citadel.model.enumeration.CitadelDescriptionRole.Name
 import utopia.citadel.model.enumeration.StandardUserRole.Owner
 import utopia.flow.time.Now
 import utopia.flow.time.TimeExtensions._
@@ -54,8 +54,7 @@ object DbOrganization
 		val membership = MembershipModel.insert(MembershipData(organizationId, founderId, Some(founderId)))
 		MemberRoleModel.insert(membership.id, Owner.id, founderId)
 		// Inserts a name for that organization
-		DbDescriptions.ofOrganizationWithId(organizationId).update(StandardDescriptionRoleId.name, languageId,
-			founderId, organizationName)
+		DbDescriptions.ofOrganizationWithId(organizationId).update(Name.id, languageId, founderId, organizationName)
 		// Returns organization id
 		organizationId
 	}